Tests/LibWeb/Text/expected/scrollable-overflow-transformed-child.txt

document scroll width: 800
positive transform scroll width: 150
